+++ grails-app/services/io/xh/hoist/telemetry/MetricsService.groovy
@@ -27,13 +27,13 @@ import io.micrometer.registry.otlp.OtlpMeterRegistry
 import io.xh.hoist.BaseService
 import io.xh.hoist.config.ConfigService
 
-import java.rmi.registry.Registry
 
 import static io.micrometer.core.instrument.config.MeterFilterReply.DENY
 import static io.micrometer.core.instrument.config.MeterFilterReply.NEUTRAL
 import static io.xh.hoist.cluster.ClusterService.instanceName
 import static io.xh.hoist.util.ClusterUtils.runOnAllInstances
 import static io.xh.hoist.util.Utils.appCode
+import static io.xh.hoist.util.Utils.prefixKeys
 
 /**
  * Central service for Micrometer metrics in a Hoist application.
@@ -233,10 +233,6 @@ class MetricsService extends BaseService {
             publishRegistry.clear()
             registry.add(publishRegistry)
         }
-    }
-
-    private static Map<String, String> prefixKeys(String prefix, Map config) {
-        (config ?: [:]).collectEntries { k, v -> ["${prefix}.${k}".toString(), v?.toString()] }
     }
 
     private String getNamespace() {