+++ src/main/groovy/io/xh/hoist/http/JSONClient.groovy
@@ -8,15 +8,18 @@
 package io.xh.hoist.http
 
 import groovy.transform.CompileStatic
+import io.opentelemetry.api.trace.Span
+import io.opentelemetry.context.Context
 import io.xh.hoist.exception.ExternalHttpException
+import io.xh.hoist.telemetry.TraceService
 import io.xh.hoist.json.JSONParser
-import io.xh.hoist.util.StringUtils
 import org.apache.hc.client5.http.impl.classic.CloseableHttpResponse
 import org.apache.hc.client5.http.classic.methods.HttpUriRequestBase
 import org.apache.hc.client5.http.impl.classic.CloseableHttpClient
 import org.apache.hc.client5.http.impl.classic.HttpClients
 
 import static io.xh.hoist.util.StringUtils.elide
+import static io.xh.hoist.util.Utils.traceService
 import static org.apache.hc.core5.http.HttpStatus.SC_NO_CONTENT
 import static org.apache.hc.core5.http.HttpStatus.SC_OK
 
@@ -120,8 +123,18 @@ class JSONClient {
             method.setHeader('Accept', 'application/json')
         }
 
+
         try {
-            ret = executeRaw(_client, method)
+            if (traceService.enabled) {
+                injectTraceContext(method)
+                def spanName = "HTTP ${method.method}".toString()
+                def tags = ['http.url': method.uri.toString(), source: 'hoist'] as Map<String, String>
+                ret = traceService.withClientSpan(spanName, tags) {
+                    executeRaw(_client, method)
+                }
+            } else {
+                ret = executeRaw(_client, method)
+            }
             statusCode = ret.code
             success = (statusCode >= SC_OK  && statusCode <= SC_NO_CONTENT)
             if (!success) {
@@ -140,6 +153,13 @@ class JSONClient {
         }
 
         return ret
+    }
+
+    private static void injectTraceContext(HttpUriRequestBase method) {
+        if (Span.current().spanContext.valid) {
+            traceService.otelSdk.propagators.textMapPropagator
+                .inject(Context.current(), method, TraceService.HTTP_SETTER)
+        }
     }
 
     protected CloseableHttpResponse executeRaw(CloseableHttpClient client, HttpUriRequestBase method) {