+++ src/main/groovy/io/xh/hoist/monitor/MonitorConfig.groovy
@@ -4,16 +4,16 @@
  *
  * Copyright © 2026 Extremely Heavy Industries Inc.
  */
-
 package io.xh.hoist.monitor
 
-import groovy.transform.MapConstructor
+import groovy.transform.InheritConstructors
+import io.xh.hoist.config.TypedConfigMap
 
 /**
  * Typed representation of `xhMonitorConfig` values.
  */
-@MapConstructor
-class MonitorConfig {
+@InheritConstructors
+class MonitorConfig extends TypedConfigMap {
     Integer monitorRefreshMins
     Integer monitorStartupDelayMins
     Integer monitorRepeatNotifyMins